requirements.txt
# 深度学习框架
torch>=1.8.0
torchvision>=0.9.0
tensorboardX>=2.4

# 数据处理和分析
numpy>=1.19.2
pandas>=1.2.4
scikit-learn>=0.24.2
Pillow>=8.2.0

# 可视化
matplotlib>=3.4.2
seaborn>=0.11.1

# 工具包
tqdm>=4.61.0
requests>=2.25.1